Update tox to run most environments in Python 3.8
The previous setup where they was run in 3.6 was based on that it
was the lowest version that was supported and if something would
work in that version it would (most probably) work in higher as
well. But on the other hand that it was the tests coverage is for,
so it does not make sense to run black et al. in a different version
than the latest.

Isorts -rc flag is apparently only recursive, not a combination of
recursive and check, so the flags was listed more explicitly.